Added dependency properties for Tile gui
Water and fertility levels were changed into dependency properties to
provide the ability to right click and display the values of the tile.
Added functionality to change the tile's water and fertility levels.
Added group boxes to selected objects on the map grid. The tile selected
groupbox is implemented, but the organism selected group box is not.