Added Animal Forced Mutation
Not a complete implementation of forced mutation. Created controls which
toggle an animal's forced mutation boolean. The animal will only begin
to mutate when the Environment seasons are activated and the simulation
begins running.
Future TODO is make the group box which toggles the forced mutation work
with both animals and plants. Currently the button only works for
Animals.