Fixed bugs with animal states
Animal states to look for plants and animals for food returned plants
and animals outside of their sense distance. This was caused by me
checking to see if their X OR Y was within sense distance as opposed to
determining the actual distance each organism was away from their food.