Plant improvements
Plants now absorb water and nutrients and use those resources to create
energy and reproduce.  Made maximum mass determined by the organism
starting mass as a child. Mad maximum energy a multiple of the mass and
an arbitrary mutate-able value. Created the dead organism which will
take the place of animals when they die.